requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Medical Imaging
monai>=1.2.0
openslide-python>=1.3.1
opencv-python-headless>=4.8.0

# Survival Analysis
scikit-survival>=0.21.0
lifelines>=0.27.7

# Data Science & Utils
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tqdm>=4.65.0
h5py>=3.9.0
jupyterlab>=4.0.0

# Hugging Face (for potential future ViT/Generative models)
transformers>=4.30.0
diffusers>=0.19.0
accelerate>=0.21.0
